cleaned up implementation based on @jenbutongit code review
