+++ designer/client/context/__tests__/FeatureFlagContext.jest.tsx
@@ -1,48 +1,27 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import { Router } from "react-router-dom";
 import { act } from "react-dom/test-utils";
-import { FeatureFlagProvider } from "../FeatureFlagContext";
-import { createMemoryHistory } from "history";
+import { FeatureFlagContext, FeatureFlagProvider } from "../FeatureFlagContext";
 import sinon from "sinon";
 import { FeatureToggleApi } from "../../api/toggleApi";
 import FeatureToggle from "../../FeatureToggle";
-import { useFeatures } from "../../hooks/featureToggling";
-const history = createMemoryHistory();
-history.push("");
-
-const customRender = (ui, { providerProps, ...renderOptions }) => {
-  const rendered = render(
-    <Router history={history}>
-      <FeatureFlagProvider value={providerProps}>{ui}</FeatureFlagProvider>
-    </Router>,
-    renderOptions
-  );
-  return {
-    ...rendered,
-    rerender: (ui, options) =>
-      customRender(ui, { container: rendered.container, ...options }),
-  };
-};
 
 describe("FeatureFlagContext", () => {
   beforeEach(() => {
     sinon.restore();
   });
 
-  const WrappingComponent = ({ children }) => {
-    return <FeatureFlagProvider>{children}</FeatureFlagProvider>;
+  const WrappingComponent = ({ value, children }) => {
+    return (
+      <FeatureFlagContext.Provider value={value}>
+        {children}
+      </FeatureFlagContext.Provider>
+    );
   };
 
   test("should show element if feature is set", async () => {
-    sinon
-      .stub(FeatureToggleApi.prototype, "fetch")
-      .callsFake(async function () {
-        return { featureEditPageDuplicateButton: true };
-      });
-
     const { findAllByText } = render(
-      <WrappingComponent>
+      <WrappingComponent value={{ featureEditPageDuplicateButton: true }}>
         <FeatureToggle feature="featureEditPageDuplicateButton">
           <button>Johnny Five Is Alive!</button>
         </FeatureToggle>
@@ -52,38 +31,24 @@ describe("FeatureFlagContext", () => {
   });
 
   test("should not show element if feature is not set", async () => {
-    const stub = sinon
-      .stub(FeatureToggleApi.prototype, "fetch")
-      .callsFake(async function () {
-        return { featureEditPageDuplicateButton: false };
-      });
-
     const { queryAllByText } = render(
-      <WrappingComponent>
+      <WrappingComponent value={{ featureEditPageDuplicateButton: false }}>
         <FeatureToggle feature="featureEditPageDuplicateButton">
           <button>Johnny Five Is Alive!</button>
         </FeatureToggle>
       </WrappingComponent>
     );
-    await act(() => stub.getCall(0).returnValue);
     expect(await queryAllByText("Johnny Five Is Alive!")).toHaveLength(0);
   });
 
   test("should not show element if feature is not defined", async () => {
-    const stub = sinon
-      .stub(FeatureToggleApi.prototype, "fetch")
-      .callsFake(async function () {
-        return { featureA: false };
-      });
-
     const { queryAllByText } = render(
-      <WrappingComponent>
+      <WrappingComponent value={{ featureA: false }}>
         <FeatureToggle feature="featureB">
           <button>Johnny Five Is Alive!</button>
         </FeatureToggle>
       </WrappingComponent>
     );
-    await act(() => stub.getCall(0).returnValue);
 
     expect(await queryAllByText("Johnny Five Is Alive!")).toHaveLength(0);
   });
@@ -96,7 +61,7 @@ describe("FeatureFlagContext", () => {
       });
 
     render(
-      <WrappingComponent>
+      <FeatureFlagProvider>
         <FeatureToggle feature="featureA">
           <button>Johnny Five Is Alive!</button>
         </FeatureToggle>
@@ -106,7 +71,7 @@ describe("FeatureFlagContext", () => {
         <FeatureToggle feature="featureC">
           <button>Johnny Five Is Alive!</button>
         </FeatureToggle>
-      </WrappingComponent>
+      </FeatureFlagProvider>
     );
     await act(() => stub.getCall(0).returnValue);
 
@@ -114,32 +79,19 @@ describe("FeatureFlagContext", () => {
   });
 
   test("should not show element if features api fails", async () => {
-    sinon.stub(FeatureToggleApi.prototype, "fetch").callsFake(function () {
-      return new Promise((_resolve, reject) => {
-        reject();
-      });
-    });
-
     const { queryAllByText } = render(
-      <WrappingComponent>
+      <WrappingComponent
+        value={() => {
+          return new Promise((_resolve, reject) => {
+            reject();
+          });
+        }}
+      >
         <FeatureToggle feature="featureEditPageDuplicateButton">
           <button>Johnny Five Is Alive!</button>
         </FeatureToggle>
       </WrappingComponent>
     );
     expect(await queryAllByText("Johnny Five Is Alive!")).toHaveLength(0);
-  });
-
-  it("should feature hook return feature context value", () => {
-    const mockContextValue = {
-      featureA: false,
-      featureB: true,
-      featureC: true,
-    };
-    sinon.stub(React, "useContext").callsFake(function () {
-      return mockContextValue;
-    });
-    const result = useFeatures();
-    expect(result).toEqual(mockContextValue);
   });
 });
